backend/requirements.txt
Django==3.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-extra-fields==3.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==10.1.0
PyJWT==2.8.0
reportlab==4.0.7